requirements.txt
asgiref==3.2.10
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cloudinary==1.22.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2
daphne==2.5.0
dj-database-url==0.5.0
Django==3.1.2
django-environ==0.4.5
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.5.0
drf-yasg==1.20.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==8.0.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.11
whitenoise==5.2.0
zope.interface==5.1.2
